change stream specs pass
