+++ spec/runners/unified/change_stream_operations.rb
@@ -10,22 +10,9 @@ module Unified
       object = entities.get_any(object_id)
       use_arguments(op) do |args|
         pipeline = args.use!('pipeline')
-        opts = {}
-        if batch_size = args.use('batchSize')
-          opts[:batch_size] = batch_size
-        end
-        if comment = args.use('comment')
-          opts[:comment] = comment
-        end
-        if full_document = args.use('fullDocument')
-          opts[:full_document] = full_document
-        end
-        if full_document_before_change = args.use('fullDocumentBeforeChange')
-          opts[:full_document_before_change] = full_document_before_change
-        end
-        if args.key?('showExpandedEvents')
-          opts[:show_expanded_events] = args.use!('showExpandedEvents')
-        end
+        opts = extract_options(args, 'batchSize', 'comment', 'fullDocument',
+          'fullDocumentBeforeChange', 'showExpandedEvents', 'timeoutMS',
+          'maxAwaitTimeMS')
         cs = object.watch(pipeline, **opts)
         if name = op.use('saveResultAsEntity')
           entities.set(:change_stream, name, cs)
@@ -37,6 +24,12 @@ module Unified
       object_id = op.use!('object')
       object = entities.get_any(object_id)
       object.try_next
+    end
+
+    def iterate_once(op)
+      stream_id = op.use!('object')
+      stream = entities.get_any(stream_id)
+      stream.try_next
     end
 
     def close(op)
